pyproject.toml
[project]
name = "hydrano"
version = "0.1.0"
description = "Lightweight Python toolkit for Cardano Hydra."
authors = [
    {name = "Nguyen Duy Khanh",email = "nguyenkhanh17112003@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9.1,<4.0.0"

[tool.poetry.dependencies]
pycardano = ">=0.15.1,<0.16.0"
websocket-client = ">=1.8.0,<2.0.0"
pyee = ">=13.0.0,<14.0.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
flake8 = "^5.0"
sphinx = "^5.0"
black = "^25.9.0"
isort = "^6.0.1"
pytest-cov = "^7.0.0"
retry = "^0.9.2"
flask = "^3.1.2"
pytest-xdist = "^3.8.0"
mypy = "^1.18.2"
sphinx-copybutton = "^0.5.2"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
